+++ testing_migrations/run.php
@@ -28,10 +28,6 @@ $configuration = [
             'password' => '123',
             'charset' => 'utf8',
         ],
-        'sqlite' => [
-            'adapter' => 'sqlite',
-            'dsn' => 'sqlite:' . __DIR__ . '/phoenix.sqlite',
-        ],
     ],
 ];