+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [Unreleased][unreleased]
 
+### [1.0.0] - 2018-06-20
 #### Changed
 - added return type (:void) for up() and down() methods in AbstractMigration (BC Break - Fix: add return type :void to all migrations)
 - replaced nette/finder with symfony/finder
@@ -177,7 +178,8 @@
 - supported methods in migrations: addColumn, addIndex, addForeignKey, dropColumn, dropIndex, dropForeignKey
 - supported column types: string, integer, boolean, text, datetime, uuid, json, char
 
-[unreleased]: https://github.com/lulco/phoenix/compare/0.13.0...HEAD
+[unreleased]: https://github.com/lulco/phoenix/compare/1.0.0...HEAD
+[1.0.0]: https://github.com/lulco/phoenix/compare/0.13.0...1.0.0
 [0.13.0]: https://github.com/lulco/phoenix/compare/0.12.1...0.13.0
 [0.12.1]: https://github.com/lulco/phoenix/compare/0.12.0...0.12.1
 [0.12.0]: https://github.com/lulco/phoenix/compare/0.11.0...0.12.0